# internal/http/testdata/openapi.golden.json
{
  "components": {
    "schemas": {
      "Actor": {
        "properties": {
          "active": {
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "display_name": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "display_name",
          "type",
          "role",
          "created_at",
          "active"
        ],
        "type": "object"
      },
...
      "CreateActorParams": {
        "properties": {
          "display_name": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "type",
          "role"
        ],
        "type": "object"
      },
...
      "Error": {
        "properties": {
          "detail": {},
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "UpdateBoardParams": {
        "properties": {
          "description": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "A human/AI collaborative task tracker with kanban boards and workflow state machines.",
    "title": "TaskFlow API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/actors": {
      "get": {
        "operationId": "actor_list",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Actor"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Success"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Client error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List all actors",
        "tags": [
          "actors"
        ]
      },
      "post": {
        "operationId": "actor_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateActorParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Actor"
                }
              }
            },
            "description": "Success"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Client error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create an actor",
        "tags": [
          "actors"
        ]
      }
    },
    "/actors/{name}": {
      "get": {
        "operationId": "actor_get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Actor"
                }
              }
            },
            "description": "Success"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Client error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get an actor by name",
        "tags": [
          "actors"
        ]
      },
      "patch": {
        "operationId": "actor_update",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateActorParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Actor"
                }
              }
            },
            "description": "Success"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Client error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update an actor",
        "tags": [
          "actors"
        ]
      }
    },
    "/admin/stats": {
      "get": {
        "operationId": "admin_stats",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemStats"
                }
              }
            },
            "description": "Success"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Client error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "System-wide statistics",
        "tags": [
          "admin"
        ]
      }
    },
    "/attachments/{id}": {
      "delete": {
        "operationId": "attachment_delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Client error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Remove an attachment",
        "tags": [
          "attachments"
        ]
      }
    },
    "/batch": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "operations": {
                    "items": {
                      "properties": {
                        "body": {
                          "description": "Request body",
                          "type": "object"
                        },
                        "idempotency_key": {
                          "description": "Per-operation idempotency key",
                          "type": "string"
                        },
                        "method": {
                          "description": "HTTP method",
                          "type": "string"
                        },
                        "path": {
                          "description": "API path",
                          "type": "string"
                        }
                      },
                      "required": [
                        "method",
                        "path"
                      ],
                      "type": "object"
                    },
                    "maxItems": 50,
                    "type": "array"
                  }
                },
                "required": [
                  "operations"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "results": {
                      "items": {
                        "properties": {
                          "body": {},
                          "status": {
                            "type": "integer"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Batch results"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Execute multiple operations in a single request (max 50)",
        "tags": [
          "batch"
        ]
      }
    },
    "/boards": {
      "get": {
        "operationId": "board_list",
        "parameters": [
          {
            "description": "Include soft-deleted boards",
            "in": "query",
            "name": "include_deleted",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Board"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Success"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Client error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List boards",
        "tags": [
          "boards"
        ]
      },
      "post": {
        "operationId": "board_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBoardParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Board"
                }
              }
            },
            "description": "Success"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Client error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create a board",
        "tags": [
          "boards"
        ]
      }
    },
    "/boards/{slug}": {
      "delete": {
        "operationId": "board_delete",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Client error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Delete a board (soft-delete)",
        "tags": [
          "boards"
        ]
      },
      "get": {
        "operationId": "board_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Board"
                }
              }
            },
            "description": "Success"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Client error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get a board",
        "tags": [
          "boards"
        ]
      },
      "patch": {
        "operationId": "board_update",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBoardParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Board"
                }
              }
            },
            "description": "Success"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Client error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update a board",
        "tags": [
          "boards"
        ]
      }
    },
    "/boards/{slug}/audit": {
      "get": {
        "operationId": "board_audit",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditEntry"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Success"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Client error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get audit log for a board",
        "tags": [
          "boards"
        ]
      }
    },
    "/boards/{slug}/detail": {
      "get": {
        "operationId": "board_detail",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardDetail"
                }
              }
            },
            "description": "Success"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Client error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get complete board with all tasks, comments, attachments, dependencies, and audit",
        "tags": [
          "boards"
        ]
      }
    },
    "/boards/{slug}/overview": {
      "get": {
        "operationId": "board_overview",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardOverview"
                }
              }
            },
            "description": "Success"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Client error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Board with task counts by state",
        "tags": [
          "boards"
        ]
      }
    },
    "/boards/{slug}/reassign": {
      "post": {
        "operationId": "board_reassign",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReassignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReassignResponse"
                }
              }
            },
            "description": "Success"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Client error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Reassign tasks to another board",
        "tags": [
          "boards"
        ]
      }
    },
    "/boards/{slug}/tags": {
      "get": {
        "operationId": "tag_list",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TagCount"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Success"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Client error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List tags in use on a board with counts",
        "tags": [
          "boards"
        ]
      }
    },
    "/boards/{slug}/tasks": {
      "get": {
        "operationId": "task_list",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by workflow state",
            "in": "query",
            "name": "state",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by assignee name",
            "in": "query",
            "name": "assignee",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by priority (critical/high/medium/low/none)",
            "in": "query",
            "name": "priority",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by tag",
            "in": "query",
            "name": "tag",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Full-text search query",
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Include tasks in terminal states",
            "in": "query",
            "name": "include_closed",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "description": "Include soft-deleted tasks",
            "in": "query",
            "name": "include_deleted",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "description": "Sort field (created_at/updated_at/priority/due_date)",
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Sort order (asc/desc)",
            "in": "query",
            "name": "order",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Success"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Client error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List tasks (with filters and search)",
        "tags": [
          "boards"
        ]
      },
      "post": {
        "operationId": "task_create",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            },
            "description": "Success"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Client error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create a task",
        "tags": [
          "boards"
        ]
      }
    },
...
    "/boards/{slug}/tasks/{num}/attachments": {
      "get": {
        "operationId": "attachment_list",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "num",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Attachment"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Success"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Client error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List attachments on a task",
        "tags": [
          "attachments"
        ]
      },
      "post": {
        "operationId": "attachment_create",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "num",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAttachmentParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Attachment"
                }
              }
            },
            "description": "Success"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Client error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Add an attachment",
        "tags": [
          "attachments"
        ]
      }
    },
...
    "/dependencies/{id}": {
      "delete": {
        "operationId": "dependency_delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Client error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Remove a dependency",
        "tags": [
          "dependencies"
        ]
      }
    },
...
  }
}